[Refactor] UserManager ui/utils 패키지로 이동
여러 패키지에서 사용되기 때문에 utils로 이동
