+++ app/src/main/java/com/and04/naturealbum/di/AuthenticationModule.kt
@@ -2,7 +2,7 @@ package com.and04.naturealbum.di
 
 import com.and04.naturealbum.data.repository.firebase.UserRepository
 import com.and04.naturealbum.ui.mypage.AuthenticationManager
-import com.and04.naturealbum.ui.mypage.UserManager
+import com.and04.naturealbum.ui.utils.UserManager
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
