+++ app/src/main/java/com/and04/naturealbum/ui/album/labelphotos/AlbumFolderViewModel.kt
@@ -8,7 +8,7 @@ import com.and04.naturealbum.data.repository.local.LocalDataRepository
 import com.and04.naturealbum.data.localdata.room.PhotoDetail
 import com.and04.naturealbum.data.model.AlbumFolderData
 import com.and04.naturealbum.ui.utils.UiState
-import com.and04.naturealbum.ui.mypage.UserManager
+import com.and04.naturealbum.ui.utils.UserManager
 import com.and04.naturealbum.utils.network.NetworkState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
